Refactor builder out as an utility
